Comparisons/SyntheticData/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --cert=None --client-cert=None --index-url=None --pip-args=None
#
build==1.3.0
    # via pip-tools
click==8.3.0
    # via pip-tools
colorama==0.4.6
    # via cook-build
cook-build==0.5.1
    # via -r requirements.in
networkx==3.5
    # via cook-build
packaging==25.0
    # via build
pip-tools==7.5.0
    # via -r requirements.in
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
wheel==0.45.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
